medical update

+++ gentechhrapplication/src/main/java/com/gentech/erp/hr/serviceimpl/MedicalEntriesServiceImpl.java
@@ -1,26 +1,19 @@
 package com.gentech.erp.hr.serviceimpl;
 
-
 import com.gentech.erp.hr.dto.MedicalEntriesDto;
 import com.gentech.erp.hr.entity.Dependant;
-import com.gentech.erp.hr.entity.Employee;
 import com.gentech.erp.hr.entity.MedicalEntries;
-import com.gentech.erp.hr.exception.ResourceNotFoundException;
-import com.gentech.erp.hr.mapper.MedicalEntriesMapper;
 import com.gentech.erp.hr.repository.DependantsRepository;
 import com.gentech.erp.hr.repository.EmployeeRepository;
 import com.gentech.erp.hr.repository.MedicalEntriesRepository;
 import com.gentech.erp.hr.service.MedicalEntriesService;
-
-
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 
 @Service
@@ -29,11 +22,14 @@ public class MedicalEntriesServiceImpl implements MedicalEntriesService {
     @Autowired
     private MedicalEntriesRepository medicalEntriesRepository;
     @Autowired
-
     private DependantsRepository dependantsRepository;
+    @Autowired
+    private EmployeeRepository employeeRepository;
+    @Autowired
+    private ModelMapper modelMapper;
 
     @Override
-    public void saveMedicalEntry(Long dependantId, MultipartFile medicalFiles, Double requestAmount) throws IOException {
+    public MedicalEntriesDto saveMedicalEntry(Long dependantId, MultipartFile medicalFiles, Double requestAmount) throws IOException {
         Dependant dependant = dependantsRepository.findById(dependantId)
                 .orElseThrow(() -> new RuntimeException("Dependant not found with ID: " + dependantId));
 
@@ -41,82 +37,41 @@ public class MedicalEntriesServiceImpl implements MedicalEntriesService {
         medicalEntry.setDependant(dependant);
         medicalEntry.setMedicalFiles(medicalFiles.getBytes());
         medicalEntry.setRequestAmount(requestAmount);
-
-
-        medicalEntriesRepository.save(medicalEntry);
+        return modelMapper.map(medicalEntriesRepository.save(medicalEntry), MedicalEntriesDto.class);
     }
 
     @Override
     public List<MedicalEntriesDto> getAllMedicalEntries() {
-        List<MedicalEntries> obj = medicalEntriesRepository.findAll();
-        List<MedicalEntriesDto> objDto = new ArrayList<>();
-        for (int i = 0; i < obj.size(); i++) {
-            objDto.add(MedicalEntriesMapper.objectToDto(obj.get(i)));
-        }
-        return objDto;
+        return medicalEntriesRepository.findAll().stream()
+                .map(medicalEntry -> modelMapper.map(medicalEntry, MedicalEntriesDto.class))
+                .toList();
     }
 
     @Override
-    public MedicalEntriesDto getMedicalEntryByMRno(Long MRno) throws Exception {
-        Optional<MedicalEntries> obj = medicalEntriesRepository.findById(MRno);
-        if (obj.isPresent()) {
-            MedicalEntriesDto dto = MedicalEntriesMapper.objectToDto(obj.get());
-            return dto;
-
-        }
-        throw new Exception("The item is not present");
+    public MedicalEntriesDto getMedicalEntryById(Long MedicalEntryId) {
+        MedicalEntries medicalEntry = medicalEntriesRepository.findById(MedicalEntryId)
+                .orElseThrow(() -> new RuntimeException("Medical Entry not found with ID: " + MedicalEntryId));
+        return modelMapper.map(medicalEntry, MedicalEntriesDto.class);
     }
-
-
 
     @Override
     public void deleteItemById(Long id) {
-        medicalEntriesRepository.deleteById(id);
-
-    }
-
-    @Autowired
-    private EmployeeRepository employeeRepository;
-
-    @Override
-    public List<MedicalEntriesDto> getMedicalEntryByeEmployeeId(Long mRno) {
-        Employee dto= employeeRepository.findById(mRno)
-                .orElseThrow(() -> new ResourceNotFoundException("Employee not found"));
-
-        List<Dependant> dtolist=dependantsRepository.findAll();
-
-        List<Dependant> sendList=new ArrayList<>();
-        for(Dependant a:dtolist)
-        {
-            if(a.getEmployee() == dto)
-            {
-                sendList.add(a);
-            }
+        if(medicalEntriesRepository.existsById(id)) {
+            medicalEntriesRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Medical Entry not found with ID: " + id);
         }
-
-
-        List<MedicalEntries> obj = medicalEntriesRepository.findAll();
-        List<MedicalEntriesDto> objDto = new ArrayList<>();
-        for (int i = 0; i < obj.size(); i++) {
-            objDto.add(MedicalEntriesMapper.objectToDto(obj.get(i)));
-        }
-
-        List<MedicalEntriesDto> finalList = new ArrayList<>();
-        for(MedicalEntriesDto b : objDto)
-        {
-            for(Dependant a:sendList)
-            {
-                if(b.getDependant().equals(a))
-                {
-                    finalList.add(b);
-                }
-            }
-        }
-        return finalList;
     }
 
     @Override
-    public void updateItem(Long dependantId, MultipartFile medicalFiles, Double requestAmount, Long medicalEntryId) throws Exception {
+    public List<MedicalEntriesDto> getMedicalEntryByeEmployeeId(Long employeeId) {
+        return medicalEntriesRepository.findAllByDependant_Employee_EmpId(employeeId).stream()
+                .map(medicalEntry -> modelMapper.map(medicalEntry, MedicalEntriesDto.class))
+                .toList();
+    }
+
+    @Override
+    public MedicalEntriesDto updateMedicalEntry(Long dependantId, MultipartFile medicalFiles, Double requestAmount, Long medicalEntryId) throws IOException {
         Dependant dependant = dependantsRepository.findById(dependantId)
                 .orElseThrow(() -> new RuntimeException("Dependant not found with ID: " + dependantId));
 
@@ -126,8 +81,7 @@ public class MedicalEntriesServiceImpl implements MedicalEntriesService {
         medicalEntry.setDependant(dependant);
         medicalEntry.setMedicalFiles(medicalFiles.getBytes());
         medicalEntry.setRequestAmount(requestAmount);
-
-        medicalEntriesRepository.save(medicalEntry);
+        return modelMapper.map(medicalEntriesRepository.save(medicalEntry), MedicalEntriesDto.class);
     }
